INTEXT-187: Allow adding advice handler on AWS outbound channel adapter JIRA: https://jira.spring.io/browse/INTEXT-187
* Modified the org/springframework/integration/aws/config/xml/spring-integration-aws-1.0.xsd to accomodate request-handler-advice-chain
	in s3 outbound adapter after feedback.
* Modified org.springframework.integration.aws.s3.config.xml.AmazonS3OutboundChannelAdapterParserTests - Added Unit Test Class with Mock Object to test the request handler chain.
* Modified s3-valid-outbound-cases.xml - Added separate definition withHandlerChain with mock Operations.
